MOVEONLY: Move from main.cpp to consensus/txverify.cpp:
-CheckFinalTx
-CheckTransaction
-Consensus::CheckTxInputs
-GetLegacySigOpCount
-GetP2SHSigOpCount
