+++ src/Makefile.am
@@ -175,6 +175,7 @@ libbitcoin_server_a_SOURCES = \
   bloom.cpp \
   chain.cpp \
   checkpoints.cpp \
+  consensus/txverify.cpp \
   init.cpp \
   leveldbwrapper.cpp \
   main.cpp \
